requirements.txt
# Minimum Django and REST framework version
Django>=1.6
djangorestframework>=2.4.3
